% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_list_null.R
\name{add_list_null}
\alias{add_list_null}
\title{Add \code{list(NULL)} to an existing list in place of absent elements}
\usage{
add_list_null(l, elm)
}
\arguments{
\item{l}{A list.}

\item{elm}{A character vector which defines list element names at which the function should check whether those elements exist and, if not, replace them with \code{list(NULL)}.}
}
\value{
A list, exactly as inputted, but in which \code{NULL} elements have been replaced by \code{list(NULL)}.
}
\description{
This function adds \code{list(NULL)} to a list at all the elements in a user-supplied vector that are \code{NULL}.
}
\examples{

l <- list(a = 10, b = 5, c = 4, d = NULL)
add_list_null(l, "d")
add_list_null(l, "e")

}
\author{
Edward Lavender
}
