% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_standard_attr.R
\name{add_standard_attr}
\alias{add_standard_attr}
\title{Add informative attributes to an object}
\usage{
add_standard_attr(
  x,
  title = deparse(substitute(x)),
  description = NULL,
  param = NULL
)
}
\arguments{
\item{x}{An object.}

\item{title}{A string providing an informative title for the object.}

\item{description}{A string providing an informative description of the object.}

\item{param}{An object (e.g., a string) containing further details about an object.}
}
\value{
The function returns the object, \code{x}, as inputted with standard attributes.
}
\description{
This function adds informative attributes to an object, including the source of the script that produced the object, a title, description and details of any parameters (if requested) and the date/time of file creation. These attributes help to maintain the links between the scripts in which objects are produced and any saved objects.
}
\details{
This function requires rstudioapi.
}
\examples{
\dontrun{
x <- runif(10, 0, 1)
out <- add_standard_attr(x)
attributes(out)
out <- add_standard_attr(x, title = "A numeric vector")
attributes(out)
}

}
\author{
Edward Lavender
}
