% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/around.R
\name{around}
\alias{around}
\title{Return the rows around a specified position in a dataframe}
\usage{
around(dataframe, position, n = 3, return = "list", verbose = TRUE)
}
\arguments{
\item{dataframe}{A dataframe.}

\item{position}{A position/numeric vector of positions around which you want to examine the rows of a dataframe.}

\item{n}{The number of adjacent positions around each position in \code{position} that you want to examine.}

\item{return}{A character specifying \code{"data.frame"} or \code{"list"}. If \code{"list"}, then all the rows around each \code{position} are returned in an element of a list; if \code{"data.frame"}, these are all returned in a single dataframe.}

\item{verbose}{A logical input that defines whether or not to print a helpful message describing the rows that are being examined.}
}
\value{
The function returns a dataframe or a list depending on the input to \code{return}.
}
\description{
This is a function like \code{\link[utils]{head}} or \code{\link[utils]{tail}} for dataframes which displays the values around a specified position.
}
\examples{

#### Example (1): A single position
# Define data
set.seed(1)
nrows <- 50
x <- runif(nrows, 1, 50)
y <- rnorm(nrows, x * 10 - 5, 60)
dat <- data.frame(x = x, y = y)
around(dat, 10, 3, "data.frame")

#### Example (2): Multiple positions, returned as a list
around(dat, c(10, 30, 39), 3, "list")

#### Example (3): Multiple positions returned as a dataframe
around(dat, c(10, 30, 39), 3, "data.frame")


}
\author{
Edward Lavender
}
