% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_dir}
\alias{check_dir}
\title{Check a directory exists}
\usage{
check_dir(arg = deparse(substitute(input)), input, check_slash = FALSE)
}
\arguments{
\item{arg}{(optional) A character string which defines the argument of a parent function.}

\item{input}{A character string which defines a directory.}

\item{check_slash}{A logical input that defines whether or not to check the end of a string for '/'. If \code{TRUE} and a '/' is lacking, this is added to the returned directory.}
}
\value{
The function checks whether or not a directory exists. If so, the function returns either the directory as inputted, or the directory with a '/' added to the end. If not, the function returns an informative error message.
}
\description{
This function checks whether a directory exists and, if not, returns an informative error message. The inputted directory can be edited with the addition of a '/' if requested.
}
\examples{
check_dir(arg = "This should work.", input = tempdir())
check_dir(arg = "Check the slash.", input = tempdir(), check_slash = TRUE)
\dontrun{
check_dir(arg = "This is silly.",
          input = ".~/this_is_a_silly_directory/")
}
}
\author{
Edward Lavender
}
