% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_if.R
\name{count_if}
\alias{count_if}
\title{Count the number of observations that meet a condition}
\usage{
count_if(condition, ...)
}
\arguments{
\item{condition}{A logical vector.}

\item{...}{Additional arguments (none currently implemented).}
}
\value{
A number which defines the number of elements in a vector which meet the condition specified.
}
\description{
This function returns the number of observations that meet a condition. This is a simple wrapper function for \code{\link[base]{which}} and \code{\link[base]{length}} designed to increase code readability and reduce code length.
}
\examples{
x <- c(1, 2, 4, 4, 3)
count_if(x < 2)

}
\author{
Edward Lavender
}
