% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stats.R
\name{dev_expl}
\alias{dev_expl}
\title{Calculate the percent deviance explained by a GLM or GAM}
\usage{
dev_expl(model)
}
\arguments{
\item{model}{A model object.}
}
\value{
The function returns the percent deviance explained as a number.
}
\description{
This function calculates the percent deviance explained by a model, such as a generalised linear model (GLM) or generalised additive model (GAM).
}
\details{
Percent deviance explained is given by \code{(model$null.deviance - model$deviance)/model$null.deviance) * 100}.
}
\examples{
n <- 100
x <- stats::runif(n, 0, 10)
y <- stats::rpois(n, x*2)
m1 <- stats::glm(y ~ x)
dev_expl(m1)
m2 <- mgcv::gam(y ~ s(x))
dev_expl(m2)

}
\author{
Edward Lavender
}
