% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_quantile_bin.R
\name{find_quantile_bin}
\alias{find_quantile_bin}
\title{Find the quantile bin in which an observation belongs}
\usage{
find_quantile_bin(
  x,
  y,
  probs = seq(0, 1, by = 0.02),
  pad_min = NULL,
  pad_max = NULL,
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of observations for which to determine the quantile bin in which they belong.}

\item{probs}{A numeric vector of probabilities, passed to \code{\link[stats]{quantile}}.}

\item{pad_min}{A number which is added to \code{x} to extend the left-hand side of the distribution of values encompassed.}

\item{pad_max}{A number which is added to \code{x} to extend the right-hand side of the distribution of values encompassed.}

\item{...}{Additional arguments passed to \code{\link[stats]{quantile}}.}
}
\value{
The function returns a dataframe with the index of the quantile bin in which each observation belongs ('bin') and the corresponding probability ('prob').
}
\description{
This function identifies the quantile interval (or 'bin') in which an observation belongs, given an empirical distribution of values (\code{x}). For observations that lie beyond the range of \code{x}, the original distribution can be 'padded' by adding an extreme minimum and/or maximum value to increase the range of the distribution (see Details). Sample quantiles are calculated for \code{x} using user-specified probabilities (\code{probs}) via \code{\link[stats]{quantile}}. The quantile bins in which each observation (\code{y}) falls is then determined and returned in a dataframe.
}
\details{
Padding is implemented by this function to express observations that are more extreme than the original distribution of observations as quantiles of that distribution. This has a small effect on the lowest/highest quantiles of the distribution but this is often unnoticeable if the quantile bins to which an observation is assigned are sufficiently large (e.g., 2 %).
}
\examples{
set.seed(0)
x <- runif(100, 0, 250)
y <- c(10, 100, 200)
find_quantile_bin(x, y)
find_quantile_bin(x, y, pad_min = 0, pad_max = 1e5)

}
\author{
Edward Lavender
}
