% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negates.R
\name{isnt.na}
\alias{isnt.na}
\title{Define available values}
\usage{
isnt.na(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[base]{NA}}.}
}
\description{
The opposite of \code{\link[base]{NA}}.
}
