% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/around.R
\name{left}
\alias{left}
\title{Select the first \code{n} columns of a dataframe or matrix}
\usage{
left(x, n = 6L)
}
\arguments{
\item{x}{A dataframe or matrix.}

\item{n}{The number of columns to return.}
}
\value{
The function returns the first \code{n} columns of a dataframe or matrix.
}
\description{
This function returns the first \code{n} columns of a dataframe or matrix.
}
\examples{
left(x = data.frame(x = 1, y = 1, z = 1), n = 4)
}
\author{
Edward Lavender
}
