% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_internal_functions.R
\name{load_internal_functions}
\alias{load_internal_functions}
\title{Load internal functions from a package}
\source{
The function was modified from code provided by Tyler Rinker (https://www.py4u.net/discuss/871790).
}
\usage{
load_internal_functions(pkg)
}
\arguments{
\item{pkg}{A character that defines the name of the package.}
}
\description{
This function loads internal functions from a package into the global environment. This is designed to support package testing.
}
\section{Warning}{

This function modifies the global environment.
}

