% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions.R
\name{pos_first_unique}
\alias{pos_first_unique}
\title{Extract the first position of each unique element in a vector}
\usage{
pos_first_unique(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
An integer vector of the positions in a vector at which each unique element appears for the first time.
}
\description{
This function extracts the position of the first appearance of each unique element in a vector. This is a simple wrapper for \code{\link[base]{which}} and \code{\link[base]{duplicated}} designed to improve code readability and reduce code length.
}
\examples{
pos_first_unique(c(1, 1, 1, 2, 4, 1, 2, 4, 9, 10, 1, 2))

}
\author{
Edward Lavender
}
