% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/around.R
\name{right}
\alias{right}
\title{Select the last \code{n} columns of a dataframe or matrix}
\usage{
right(x, n = 6L)
}
\arguments{
\item{x}{A dataframe or matrix.}

\item{n}{The number of columns to return.}
}
\value{
The function returns the last \code{n} columns of a dataframe or matrix.
}
\description{
This function returns the last \code{n} columns of a dataframe or matrix.
}
\examples{
right(x = data.frame(x = 1, y = 1, z = 1), n = 2L)
}
\author{
Edward Lavender
}
