% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_range.R
\name{round_range}
\alias{round_range}
\title{Round a range}
\usage{
round_range(x, digits = "auto")
}
\arguments{
\item{x}{A vector of two numbers to be round down and up respectively.}

\item{digits}{The precision of rounding (e.g. to the nearest 0.1). "auto" is allowed which attempts to define suitable accuracy automatically.}
}
\value{
A number.
}
\description{
This rounds the lower and upper values of a range down and up, respectively, by a user-specified amount. This can be useful, for example, in the creation of plots with suitable axis limits. The function uses \code{\link[plyr]{round_any}}.
}
\examples{
round_range(c(0.987, 10.982), "auto")

}
\author{
Edward Lavender
}
