% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_range.R
\name{seq_range}
\alias{seq_range}
\title{Define a sequence between a range of numbers}
\usage{
seq_range(range, ...)
}
\arguments{
\item{range}{A vector of two numbers representing \code{from} and \code{to} in \code{seq()}.}

\item{...}{Other arguments passed to \code{\link[base]{seq}}.}
}
\description{
This function is an implementation of \code{\link[base]{seq}} in which the user can simply input a range (i.e., a vector of two numbers).
}
\examples{

#### Define a vector
x <- runif(10, 1, 5)

#### Example (1) A sequence of specified length
seq_range(range(x), length.out = 10)

#### Example (2) A sequence with a specified increment
seq_range(range(x), by = 0.001)

}
\author{
Edward Lavender
}
