% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sterm.R
\name{sterm}
\alias{sterm}
\title{The names of \code{mgcv} smooth terms}
\usage{
sterm(s = "ti", vars = c("hourofday", "julian_day"))
}
\arguments{
\item{s}{A character vector specifying the type of smooth term (e.g. "s", "te", "ti", "crs", "cr" etc.)}

\item{vars}{A character vector or list specifying variable names.}
}
\description{
This function defines the names of \code{mgcv} smooth terms based on the type of smooth and the variable name(s). This is useful, for instance, for extracting term-specific predictions from a \code{\link[mgcv]{predict.gam}} or \code{\link[mgcv]{predict.bam}} object.
}
\examples{
# Supplying variables as a list:
sterm(s = "s", vars = list("th_interp"))
sterm(s = "ti", vars = list("hourofday", "julian_day"))
# Supplying variables as character vectors:
sterm(s = "s", vars = "th_interp")
sterm(s = "s", vars = c("hourofday", "julian_day"))

}
\author{
Edward Lavender
}
