% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substr_end.R
\name{substr_end}
\alias{substr_end}
\title{Extract the last n letters of a character}
\usage{
substr_end(x, n)
}
\arguments{
\item{x}{A character or character vector.}

\item{n}{A number which defines the number of letters from the end of the word to be retained.}
}
\value{
A character or character vector, as inputted, but including only the last n letters of each element.
}
\description{
This is a wrapper for \code{\link[base]{substr}} for the extraction of the last n letters of a character/word.
}
\examples{

substr_end(c("edward", "john"), 2)
substr_end(as.character(seq.Date(as.Date("2016-01-01"), as.Date("2016-01-31"), 1)), 5)

}
\author{
Edward Lavender
}
