% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_coef.R
\name{tidy_coef}
\alias{tidy_coef}
\title{Tidy a table of coefficients}
\usage{
tidy_coef(
  coef,
  coef_names = rownames(coef),
  col_names = c("Coefficient", "Estimate", "SE", "z-value", "p-value")[1:(ncol(coef) +
    1)],
  f = function(x) prettyGraphics::add_lagging_point_zero(round(x, digits = 3), n = 3),
  f_index = 1:ncol(coef),
  file = NULL,
  quote = FALSE,
  sep = ",",
  row.names = FALSE,
  ...
)
}
\arguments{
\item{coef}{A matrix of model coefficients, usually from \code{\link[stats]{coef}}. The matrix is expected to have row names (coefficient names) and numeric coefficients.}

\item{coef_names}{(optional) A character vector of 'tidy' coefficient names.}

\item{col_names}{(optional) A character vector of 'tidy' column names. The first column name should be the name of the column containing the coefficient names.}

\item{f}{(optional) A function to be applied to specified columns to tidy numeric outputs.}

\item{f_index}{An integer vector of the indices of the columns in \code{coef} to which the function \code{f} is applied. By default, if \code{f} is supplied, it is applied to all columns.}

\item{file}{(optional) A character which defines the name of the file to be saved. This is passed to \code{\link[utils]{write.table}}.}

\item{quote}{A logical value which defines whether or not to surround character or factor columns with quotes. This is passed to \code{\link[utils]{write.table}}.}

\item{sep}{The field separator string. This is passed to \code{\link[utils]{write.table}}.}

\item{row.names}{A logical value which defines whether or not row names are to be saved (or a character vector or the row names to be written). This is passed to \code{\link[utils]{write.table}}.}

\item{...}{Other arguments passed to \code{\link[utils]{write.table}}.}
}
\value{
The function returns/saves a tidier table of coefficients.
}
\description{
This function tidies a table of coefficients. Values can be processed (e.g., by rounding) and row and column names adjusted. A text file can be saved that can be imported into other programmes (e.g., Microsoft Word).
}
\examples{
n <- 100
x <- runif(n, 0, 1)
y <- rnorm(n, 5*x, 10)
mod <- lm(y ~ x)
tidy_coef(coef(summary(mod)))
}
\author{
Edward Lavender
}
