% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.add.R
\docType{package}
\name{utils.add}
\alias{utils.add}
\title{Additional Utilities for Base R}
\description{
utils.add provides functions for an eclectic and evolving collection of R operations implemented by the author in the course of scientific research. This includes functions which facilitate data extraction, learning from data, data modification and common mathematical functions.
}
\author{
Edward Lavender
}
